fix(Hotfix): SR-FULL 지적 H1~H4 적용
- H1: PoolingSubsystem::ActorPool UPROPERTY 추가, 키 TObjectPtr<UClass>로 강화
- H2: GA_Base::SPAWN_OFFSET → EditDefaultsOnly UPROPERTY (4회차 하드코딩 교정)
- H3: GA_ProjectileAttack constexpr int32→float 타입 불일치 수정
- H4: DataTableStructs Arc/Area/Defense 구조체 기본값 미설정 필드 추가

+++ Source/RoastStaffGAS/Private/GAS/Abilities/GA_ProjectileAttack.cpp
@@ -233,7 +233,7 @@ bool UGA_ProjectileAttack::HandleArcType(FProjectileInitData& OutInitData, const
 		return false;
 	}
 	
-	constexpr int32 LAUNCH_ANGLE_CLAMP = 80.f;
+	constexpr float LAUNCH_ANGLE_CLAMP = 80.f;
 	OutInitData.LaunchAngle  = FMath::Clamp(ArcData.LaunchAngle, -LAUNCH_ANGLE_CLAMP, LAUNCH_ANGLE_CLAMP);
 	OutInitData.GravityScale = ArcData.GravityScale;
 	return true;
